Found and moved NoopBackgroundEvent to events sub-package
