+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/events/NoopBackgroundEvent.java
@@ -14,24 +14,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.clients.consumer.internals;
+package org.apache.kafka.clients.consumer.internals.events;
 
-import org.apache.kafka.clients.consumer.internals.events.BackgroundEvent;
+import java.util.Objects;
 
 /**
- * Noop event. Intentionally left it here for demonstration purpose.
+ * No-op event. Intentionally left it here for demonstration purpose.
  */
 public class NoopBackgroundEvent extends BackgroundEvent {
 
-    public final String message;
+    private final String message;
 
     public NoopBackgroundEvent(final String message) {
         super(Type.NOOP);
-        this.message = message;
+        this.message = Objects.requireNonNull(message);
+    }
+
+    public String message() {
+        return message;
     }
 
     @Override
     public String toString() {
-        return getClass() + "_" + this.message;
+        return "NoopBackgroundEvent{" +
+                "message='" + message + '\'' +
+                ", type=" + type +
+                '}';
     }
 }
